Move planner imports to import directory
This change moves `planner_import.c` to the `import` subdirectory and
renames it to `planner.c`. The code can now be included as
`import/planner.h`.
